SpreadsheetMetadataLoader extracted SpreadsheetMetadataContext
- Closes https://github.com/mP1/walkingkooka-spreadsheet/issues/9304
- Extract SpreadsheetMetadataLoader
